Remove CPI API v1 Support
Phase 1: Simplify ExternalCpiResponseWrapper by removing v1 branches.
- create_vm: no longer wraps a bare string result in an array; v2+ CPIs
  always return [vm_cid, network_settings] directly
- attach_disk: no longer returns nil for v1; always validates and returns
  the disk hint
- create_stemcell: condense conditional to a single expression

Phase 2: dummy_cpi now unconditionally uses DummyV2. The case statement
that dispatched on the requested api_version is removed; v1 was the only
other branch and it is no longer supported.

Now that CPI v1 is no longer tested, the Dummy/DummyV2 inheritance
hierarchy is unnecessary. Combine them into a single top-level
DummyCpi class with v2 behaviour baked in:

Fail closed when CPI does not report api_version via info; fix v2 test shapes

Phase 3: Remove all v1 test stubs and fixtures.
- Delete spec/integration/cpi_versions/director_cpi_v1_spec.rb entirely
- Remove 'when CPI is v1' context from unmanaged_persistent_disks_spec.rb
- Change sandbox default dummy_cpi_api_version from 1 to 2
- Remove 'when cpi_version is 1' describe block from
  external_cpi_response_wrapper_spec.rb
- Update preferred_cpi_api_version and cpi_api_version stubs from 1 to 2
  in: create_vm_step_spec, cloud_factory_spec, az_cloud_factory_spec,
  package_compile_stage_spec, unresponsive_agent_spec,
  update_stage_functional_spec, compilation_instance_pool_spec
- Fix pre-existing bug in external_cpi_spec.rb: replace undefined
  asset_path() helper call with an explicit File.expand_path relative to
  __dir__, pointing at spec/assets/bin/dummy_cpi.
- Update the mocks to return [vm_cid, network_settings] tuples as a v2
  CPI would, keeping the behaviour consistent with the new wrapper code.
- Merge Dummy and DummyV2 into DummyCpi, drop Bosh::Clouds namespace
