+++ src/bosh-director/lib/clouds/external_cpi_response_wrapper.rb
@@ -10,6 +10,7 @@ module Bosh::Clouds
 
     def current_vm_id(*arguments); invoke_cpi_method(__method__.to_s, *arguments); end
     def delete_stemcell(*arguments); invoke_cpi_method(__method__.to_s, *arguments); end
+    def create_vm(*arguments); invoke_cpi_method(__method__.to_s, *arguments); end
     def delete_vm(*arguments); invoke_cpi_method(__method__.to_s, *arguments); end
     def create_network(*arguments); invoke_cpi_method(__method__.to_s, *arguments); end
     def delete_network(*arguments); invoke_cpi_method(__method__.to_s, *arguments); end
@@ -20,7 +21,6 @@ module Bosh::Clouds
     def create_disk(*arguments); invoke_cpi_method(__method__.to_s, *arguments); end
     def has_disk(*arguments); invoke_cpi_method(__method__.to_s, *arguments); end
     def delete_disk(*arguments); invoke_cpi_method(__method__.to_s, *arguments); end
-    def attach_disk(*arguments); invoke_cpi_method(__method__.to_s, *arguments); end
     def detach_disk(*arguments); invoke_cpi_method(__method__.to_s, *arguments); end
     def snapshot_disk(*arguments); invoke_cpi_method(__method__.to_s, *arguments); end
     def delete_snapshot(*arguments); invoke_cpi_method(__method__.to_s, *arguments); end
@@ -31,49 +31,30 @@ module Bosh::Clouds
     def calculate_vm_cloud_properties(*arguments); invoke_cpi_method(__method__.to_s, *arguments); end
     def info; invoke_cpi_method(__method__.to_s); end
 
-    def invoke_cpi_method(method, *arguments)
-      @cpi.public_send(method, *arguments)
+    def create_stemcell(*args)
+      final_args = @cpi_api_version >= 3 ? args : args.take(2)
+      @cpi.create_stemcell(*final_args)
     end
-
-    def create_vm(*args)
-      cpi_response = @cpi.create_vm(*args)
-
-      response = []
-      if @cpi_api_version >= 2
-        response = cpi_response
-      else
-        response << cpi_response
-      end
-
-      response
-    end
-
-
-    def create_stemcell(*args) 
-      final_args = args.take(2)
-      if @cpi_api_version >= 3
-        final_args = args
-      end
-      return  @cpi.create_stemcell(*final_args)
-    end
-
 
     def attach_disk(*args)
       cpi_response = @cpi.attach_disk(*args)
-
-      if @cpi_api_version >= 2
-        raise Bosh::Clouds::AttachDiskResponseError, 'No disk_hint' if cpi_response.nil? || cpi_response.empty?
-        cpi_response
-      else
-        nil
-      end
+      raise Bosh::Clouds::AttachDiskResponseError, 'No disk_hint' if cpi_response.nil? || cpi_response.empty?
+      cpi_response
     end
 
     private
 
+    def invoke_cpi_method(method, *arguments)
+      @cpi.public_send(method, *arguments)
+    end
+
     def check_cpi_api_support
-      unsupported = @cpi_api_version > Bosh::Director::Config.preferred_cpi_api_version
-      raise Bosh::Clouds::NotSupported, "CPI API version #{@cpi_api_version} is not supported." if unsupported
+      if @cpi_api_version < 2
+        raise Bosh::Clouds::NotSupported, "CPI API version #{@cpi_api_version} is not supported. Minimum required version is 2."
+      end
+      if @cpi_api_version > Bosh::Director::Config.preferred_cpi_api_version
+        raise Bosh::Clouds::NotSupported, "CPI API version #{@cpi_api_version} is not supported."
+      end
     end
   end
 end
