+++ src/bosh-director/spec/unit/bosh/director/problem_handlers/missing_vm_spec.rb
@@ -110,7 +110,7 @@ module Bosh::Director
       before do
         allow(Config).to receive(:uuid).and_return('woof-uuid')
         allow(Config).to receive(:cloud_options).and_return('provider' => { 'path' => '/path/to/default/cpi' })
-        allow(fake_cloud).to receive(:info)
+        allow(fake_cloud).to receive(:info).and_return({ 'api_version' => 2 })
         allow(fake_cloud).to receive(:set_vm_metadata)
         allow(fake_cloud).to receive(:request_cpi_api_version=)
         allow(fake_cloud).to receive(:request_cpi_api_version)
@@ -165,7 +165,7 @@ module Bosh::Director
               'bosh' => { 'group' => String, 'groups' => anything },
             },
           )
-          .and_return('new-vm-cid')
+          .and_return(['new-vm-cid', {}])
 
         fake_job_context
