+++ src/bosh-director/spec/unit/bosh/director/problem_handlers/unresponsive_agent_spec.rb
@@ -54,8 +54,8 @@ module Bosh::Director
       allow(Config).to receive(:nats_client_ca_private_key_path).and_return(director_config['nats']['client_ca_private_key_path'])
       allow(Config).to receive(:nats_client_ca_certificate_path).and_return(director_config['nats']['client_ca_certificate_path'])
 
-      allow(cloud).to receive(:info)
-      allow(cloud).to receive(:request_cpi_api_version).and_return(1)
+      allow(cloud).to receive(:info).and_return('api_version' => 2)
+      allow(cloud).to receive(:request_cpi_api_version).and_return(2)
       allow(cloud).to receive(:request_cpi_api_version=)
       allow(Bosh::Clouds::ExternalCpi).to receive(:new).with('/path/to/default/cpi',
                                                              'woof-uuid',
@@ -220,7 +220,7 @@ module Bosh::Director
             networks,
             [],
             { 'key1' => 'value1', 'bosh' => { 'group' => String, 'groups' => anything } }
-          ).and_return('new-vm-cid')
+          ).and_return(['new-vm-cid', {}])
 
           expect(fake_new_agent).to receive(:wait_until_ready).ordered
           expect(fake_new_agent).to receive(:update_settings).ordered
