+++ src/bosh-director/spec/unit/clouds/external_cpi_response_wrapper_spec.rb
@@ -561,151 +561,4 @@ describe Bosh::Clouds::ExternalCpiResponseWrapper do
     end
   end
 
-  describe 'when cpi_version is 1' do
-    let(:cpi_api_version) { 1 }
-
-    describe '#create_vm' do
-      let(:cpi_response) { JSON.dump(result: 'fake-result', error: nil, log: 'fake-log') }
-      let(:redacted_network_settings) { nil }
-      let(:expected_response) { ['fake-result'] }
-
-      it_calls_cpi_method(
-        :create_vm,
-        'fake-agent-id',
-        'fake-stemcell-cid',
-        { 'cloud' => 'props' },
-        nil,
-        ['fake-disk-cid'],
-        'bosh' => {
-          'group' => 'my-group',
-          'groups' => ['my-first-group'],
-          'password' => 'my-secret-password',
-        },
-        'other' => 'value',
-      )
-
-      context 'when network settings hash cloud properties is absent' do
-        let(:expected_response) { ['fake-result'] }
-        let(:redacted_network_settings) do
-          {
-            'net' => {
-              'type' => 'manual',
-              'ip' => '10.10.0.2',
-              'netmask' => '255.255.255.0',
-              'default' => %w[dns gateway],
-              'dns' => ['10.10.0.2'],
-              'gateway' => '10.10.0.1',
-            },
-          }
-        end
-
-        it_calls_cpi_method(
-          :create_vm,
-          'fake-agent-id',
-          'fake-stemcell-cid',
-          { 'cloud' => 'props' },
-          {
-            'net' => {
-              'type' => 'manual',
-              'ip' => '10.10.0.2',
-              'netmask' => '255.255.255.0',
-              'default' => %w[dns gateway],
-              'dns' => ['10.10.0.2'],
-              'gateway' => '10.10.0.1',
-            },
-          },
-          ['fake-disk-cid'],
-          'bosh' => {
-            'group' => 'my-group',
-            'groups' => ['my-first-group'],
-            'password' => 'my-secret-password',
-          },
-          'other' => 'value',
-        )
-      end
-    end
-
-    describe('#attach_disk') do
-      let(:cpi_response) { JSON.dump(result: 'fake-result', error: nil, log: 'fake-log') }
-      let(:expected_response) { nil }
-      it_calls_cpi_method(:attach_disk, 'fake-vm-cid', 'fake-disk-cid')
-    end
-
-    describe 'forwards all other methods without change' do
-      let(:cpi_response) { JSON.dump(result: 'fake-result', error: nil, log: 'fake-log') }
-      let(:expected_response) { 'fake-result' }
-
-      context('#current_vm_id') do
-        it_calls_cpi_method(:current_vm_id)
-      end
-
-      context('#create_stemcell') do
-        it_calls_cpi_method(:create_stemcell, 'fake-stemcell-image-path', 'cloud' => 'props')
-      end
-
-      context('#delete_stemcell') do
-        it_calls_cpi_method(:delete_stemcell, 'fake-stemcell-cid')
-      end
-
-      context('#delete_vm') do
-        it_calls_cpi_method(:delete_vm, 'fake-vm-cid')
-      end
-
-      context('#has_vm') do
-        it_calls_cpi_method(:has_vm, 'fake-vm-cid')
-      end
-
-      context('#reboot_vm') do
-        it_calls_cpi_method(:reboot_vm, 'fake-vm-cid')
-      end
-
-      context('#set_vm_metadata') do
-        it_calls_cpi_method(:set_vm_metadata, 'fake-vm-cid', 'metadata' => 'hash')
-      end
-
-      context('#set_disk_metadata') do
-        it_calls_cpi_method(:set_disk_metadata, 'fake-disk-cid', 'metadata' => 'hash')
-      end
-
-      context('#create_disk') do
-        it_calls_cpi_method(:create_disk, 100_000, { 'type' => 'gp2' }, 'fake-vm-cid')
-      end
-
-      context('#has_disk') do
-        it_calls_cpi_method(:has_disk, 'fake-disk-cid')
-      end
-
-      context('#delete_disk') do
-        it_calls_cpi_method(:delete_disk, 'fake-disk-cid')
-      end
-
-      context('#detach_disk') do
-        it_calls_cpi_method(:detach_disk, 'fake-vm-cid', 'fake-disk-cid')
-      end
-
-      context('#snapshot_disk') do
-        it_calls_cpi_method(:snapshot_disk, 'fake-disk-cid')
-      end
-
-      context('#delete_snapshot') do
-        it_calls_cpi_method(:delete_snapshot, 'fake-snapshot-cid')
-      end
-
-      context('#resize_disk') do
-        it_calls_cpi_method(:resize_disk, 'fake-disk-cid', 1024)
-      end
-
-      context('#get_disks') do
-        it_calls_cpi_method(:get_disks, 'fake-vm-cid')
-      end
-
-      context('#ping') do
-        it_calls_cpi_method(:ping)
-      end
-
-      context('#info') do
-        it_calls_cpi_method(:info)
-      end
-    end
-  end
 end
