+++ src/bosh-director/spec/unit/clouds/external_cpi_spec.rb
@@ -649,12 +649,13 @@ describe Bosh::Clouds::ExternalCpi do
     let(:logger) { Logging::Logger.new('ExternalCpi') }
     let(:cpi_log_path) { '/var/vcap/task/5/cpi' }
     let(:config) { double('Bosh::Director::Config', logger: logger, cpi_task_log: cpi_log_path) }
-    let(:external_cpi) { described_class.new(asset_path("bin/dummy_cpi"), 'fake-director-uuid', logger) }
+    let(:external_cpi_executable) { File.expand_path('../../assets/bin/dummy_cpi', __dir__) }
+    let(:external_cpi) { described_class.new(external_cpi_executable, 'fake-director-uuid', logger) }
 
     before do
       stub_const('Bosh::Director::Config', config)
       FileUtils.mkdir_p('/var/vcap/task/5')
-      allow(File).to receive(:executable?).with(asset_path('bin/dummy_cpi')).and_return(true)
+      allow(File).to receive(:executable?).with(external_cpi_executable).and_return(true)
       allow(Open3).to receive(:popen3).and_wrap_original do |original_method, *args, &block|
         # We need to make sure Open3.popen3 gets called with a env path where ruby exists.
         # We happen to know it is /usr/local/bin/ in this case.
