+++ src/spec/integration/cpi_and_agent_spec.rb
@@ -10,7 +10,7 @@ describe 'CPI and Agent:', type: :integration do
         target: 'cpi',
         method: 'create_vm',
         agent_id: agent_id,
-        response_matcher: be(vm_cid),
+        response_matcher: eq([vm_cid, {}]),
       },
       { target: 'cpi', method: 'set_vm_metadata', vm_cid: vm_cid },
       { target: 'agent', method: 'ping', agent_id: agent_id, can_repeat: true },
@@ -63,6 +63,8 @@ describe 'CPI and Agent:', type: :integration do
   def attach_disk_sequence(vm_cid, agent_id, disk_cid)
     [
       { target: 'cpi', method: 'attach_disk', vm_cid: vm_cid, argument_matcher: include('disk_id' => disk_cid) },
+      { target: 'agent', method: 'ping', agent_id: agent_id, can_repeat: true },
+      { target: 'agent', method: 'add_persistent_disk', agent_id: agent_id },
       { target: 'cpi', method: 'set_disk_metadata', argument_matcher: include('disk_cid' => disk_cid) },
       { target: 'agent', method: 'ping', agent_id: agent_id, can_repeat: true },
       { target: 'agent', method: 'mount_disk', agent_id: agent_id, argument_matcher: match([disk_cid]) },
@@ -196,7 +198,8 @@ describe 'CPI and Agent:', type: :integration do
     let(:instances) { 1 }
 
     let(:old_vm_id) do
-      old_vm_create_invocation.response
+      response = old_vm_create_invocation.response
+      response.is_a?(Array) ? response.first : response
     end
 
     let(:old_vm_agent_id) do
