+++ src/spec/integration_support/bin/dummy_cpi
@@ -8,8 +8,7 @@ require 'stringio'
 require 'yaml'
 require 'cloud'
 
-require_relative '../clouds/dummy'
-require_relative '../clouds/dummy_v2'
+require_relative '../dummy_cpi'
 
 module Logging
   def self.appenders
@@ -65,12 +64,11 @@ begin
   context = request['context']
   requested_api_version = request['api_version']
 
-  dummy = case requested_api_version
-          when 2
-            Bosh::Clouds::DummyV2.new(cpi_config, context)
-          else
-            Bosh::Clouds::Dummy.new(cpi_config, context, cpi_config['api_version'])
-          end
+  unless requested_api_version.nil? || requested_api_version == 2
+    raise "Unsupported CPI API version: #{requested_api_version}. Only version 2 is supported."
+  end
+
+  dummy = DummyCpi.new(cpi_config, context)
 
   result = dummy.send(command, *arguments)
 rescue StandardError, ScriptError => e
