+++ src/spec/integration_support/sandbox.rb
@@ -6,7 +6,7 @@ require 'securerandom'
 require 'tmpdir'
 require 'tempfile'
 
-require 'integration_support/clouds/dummy'
+require 'integration_support/dummy_cpi'
 
 require 'integration_support/constants'
 require 'integration_support/service'
@@ -269,7 +269,7 @@ module IntegrationSupport
 
       # Note that this is not the same object
       # as dummy cpi used as the external CPI subprocess
-      @cpi = Bosh::Clouds::Dummy.new(
+      @cpi = DummyCpi.new(
         {
           'dir' => cloud_storage_dir,
           'agent' => {
@@ -283,8 +283,7 @@ module IntegrationSupport
           'nats' => @nats_url,
           'log_buffer' => @logger,
         },
-        {},
-        1
+        {}
       )
 
       reconfigure
@@ -472,7 +471,7 @@ module IntegrationSupport
       @agent_wait_timeout = options.fetch(:agent_wait_timeout, 600)
       @keep_unreachable_vms = options.fetch(:keep_unreachable_vms, false)
       @with_incorrect_nats_server_ca = options.fetch(:with_incorrect_nats_server_ca, false)
-      @dummy_cpi_api_version = options.fetch(:dummy_cpi_api_version, 1)
+      @dummy_cpi_api_version = options.fetch(:dummy_cpi_api_version, 2)
       @nats_url = "nats://localhost:#{nats_port}"
       @cpi.options['nats'] = @nats_url
